junqueiracrew-analisefii/requirements.txt
uvicorn
crewai
typing
weasyprint
jinja2
fastapi
python-dotenv
pandas
